Safari-609.1.11.1/LayoutTests/inspector/heap/garbageCollected-expected.txt

Test for the Heap.garbageCollected event.


== Running test suite: Heap.garbageCollected
-- Running test case: TriggerGCs
PASS: Should see both a Partial and Full Garbage Collection.
